requirements.txt
kubernetes==28.1.0
PyYAML==6.0.1
testcontainers==3.7.1
docker==7.0.0